streaming-demo/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
sse-starlette==1.8.2
python-dotenv==1.0.0
httpx==0.25.2 